requirements.txt
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
stripe==5.4.0
tomli==2.0.1
urllib3==1.26.15
yapf==0.33.0
